Update build.gradle for build errors
the project will no longer build with buildToolsVersion "23.0.0", android studio will automatically update this to "25.0.0" for you if you allow it but people are having an issues with this as stated in issue #47 https://github.com/youtube/yt-watchme/issues/47 of the master branch.  Updating the source code to "25.0.0" should resolve this error.